Moved device memory resource files to mr/device folder.
