+++ include/rmm/detail/memory_manager.hpp
@@ -31,7 +31,7 @@
 #include <set>
 #include <mutex>
 
-#include <rmm/mr/device_memory_resource.hpp>
+#include <rmm/mr/device/device_memory_resource.hpp>
 
 #include <rmm/rmm_api.h>
 #include <rmm/detail/cnmem.h>
