+++ include/rmm/device_buffer.hpp
@@ -15,8 +15,8 @@
  */
 #pragma once
 
-#include <rmm/mr/default_memory_resource.hpp>
-#include <rmm/mr/device_memory_resource.hpp>
+#include <rmm/mr/device/default_memory_resource.hpp>
+#include <rmm/mr/device/device_memory_resource.hpp>
 
 #include <cuda_runtime_api.h>
 #include <cassert>
