+++ src/rmm.cpp
@@ -23,10 +23,10 @@
 
 #include "rmm/rmm.h"
 #include "rmm/detail/memory_manager.hpp"
-#include "rmm/mr/cnmem_memory_resource.hpp"
-#include "rmm/mr/cnmem_managed_memory_resource.hpp"
-#include "rmm/mr/managed_memory_resource.hpp"
-#include "rmm/mr/cuda_memory_resource.hpp"
+#include "rmm/mr/device/cnmem_memory_resource.hpp"
+#include "rmm/mr/device/cnmem_managed_memory_resource.hpp"
+#include "rmm/mr/device/managed_memory_resource.hpp"
+#include "rmm/mr/device/cuda_memory_resource.hpp"
