+++ tests/device_buffer_tests.cpp
@@ -17,12 +17,12 @@
 #include "gtest/gtest.h"
 
 #include <rmm/device_buffer.hpp>
-#include <rmm/mr/cnmem_memory_resource.hpp>
-#include <rmm/mr/cnmem_managed_memory_resource.hpp>
-#include <rmm/mr/cuda_memory_resource.hpp>
-#include <rmm/mr/default_memory_resource.hpp>
-#include <rmm/mr/device_memory_resource.hpp>
-#include <rmm/mr/managed_memory_resource.hpp>
+#include <rmm/mr/device/cnmem_memory_resource.hpp>
+#include <rmm/mr/device/cnmem_managed_memory_resource.hpp>
+#include <rmm/mr/device/cuda_memory_resource.hpp>
+#include <rmm/mr/device/default_memory_resource.hpp>
+#include <rmm/mr/device/device_memory_resource.hpp>
+#include <rmm/mr/device/managed_memory_resource.hpp>
 
 #include <cuda_runtime_api.h>
 #include <cstddef>
