+++ tests/device_scalar_tests.cpp
@@ -17,8 +17,8 @@
 #include "gtest/gtest.h"
 
 #include <rmm/device_scalar.hpp>
-#include <rmm/mr/default_memory_resource.hpp>
-#include <rmm/mr/device_memory_resource.hpp>
+#include <rmm/mr/device/default_memory_resource.hpp>
+#include <rmm/mr/device/device_memory_resource.hpp>
 
 #include <cuda_runtime_api.h>
 #include <chrono>
